Unbreak php 5.6
Ok so that's really funny. There were secretly a LOT of errors in our tests, but they were passing since the previous code was forcing a warning instead of fatal error.

PHP 5.6 doesn't support function type hinting; so the fn(string $thing = '') was causing it to treat string as a class, which is supported. But once the type hint was removed it was looking for a class that doesn't exist.